+++ ideemyouworthy/playlistmanager.py
@@ -20,7 +20,7 @@ class PlaylistManager:
         if not self.master_playlist_file.exists():
             self.master_playlist_file.touch()
             self.master_playlist_file.write_text(json.dumps({}))
-            self.logger.log("Created master playlist file")
+            self.logger.info("Created master playlist file")
 
         self.custom_playlist_file = Path(Path.cwd().parents[0] / "custom_playlists.json")
         if not self.custom_playlist_file.exists():
@@ -28,8 +28,8 @@ class PlaylistManager:
             sample_custom_playlists = collections.OrderedDict()
             sample_custom_playlists["https://open.spotify.com/playlist/3p22aU2NEvY8KErZAoWSJD"] = "[example--will not be used]"
             sample_custom_playlists["https://open.spotify.com/playlist/2JD9j9iGtPGaupLf0NwZXe"] = "[example--will not be used]"
-            self.custom_playlist_file.write_text(json.dumps(sample_custom_playlists, indent = 4))
-            self.logger.log("Created custom playlist file")
+            self.custom_playlist_file.write_text(json.dumps(sample_custom_playlists, indent = 4, ensure_ascii = False))
+            self.logger.info("Created custom playlist file")
 
     def uri_to_url(self, uri):
         uri_array = uri.split(":")
@@ -46,7 +46,7 @@ class PlaylistManager:
             return None
 
     def get_new_user_playlists(self):
-        self.logger.log("Starting user playlist retrieval")
+        self.logger.info("Starting user playlist retrieval")
         spotify_username = self.account_manager.account_info_dict["SPOTIFY_USERNAME"]
         playlists = self.spotify_manager.user_playlists(spotify_username)
         new_playlists = collections.OrderedDict()
@@ -58,7 +58,7 @@ class PlaylistManager:
             else:
                 playlists = None
 
-        self.logger.log("Finished user playlist retrieval")
+        self.logger.info("Finished user playlist retrieval")
         return new_playlists
 
     def create_playlist_files(self, new_playlists):
@@ -72,13 +72,13 @@ class PlaylistManager:
                 file_path.write_text(json.dumps({}))
                 new_file_count += 1
 
-        self.logger.log("Created " + str(new_file_count) + " new playlist files")
+        self.logger.info("Created " + str(new_file_count) + " new playlist files")
 
     def store_user_playlists(self, new_playlists):
-        self.master_playlist_file.write_text(json.dumps(new_playlists, indent = 4))
+        self.master_playlist_file.write_text(json.dumps(new_playlists, indent = 4, ensure_ascii = False))
 
     def store_custom_playlists(self, new_playlists):
-        self.custom_playlist_file.write_text(json.dumps(new_playlists, indent = 4))
+        self.custom_playlist_file.write_text(json.dumps(new_playlists, indent = 4, ensure_ascii = False))
 
     def read_custom_playlists(self):
         custom_playlists_dict = json.loads(self.custom_playlist_file.read_text())
@@ -98,7 +98,7 @@ class PlaylistManager:
             if not custom_playlists_dict[playlist] == "[example--will not be used]":
                 custom_playlists_dict[playlist] = playlist_object["name"]
 
-        self.logger.log("Read custom playlists")
+        self.logger.info("Read custom playlists")
 
         self.store_custom_playlists(custom_playlists_dict)
 
@@ -108,13 +108,13 @@ class PlaylistManager:
         custom_playlists_dict = json.loads(self.custom_playlist_file.read_text())
         custom_playlists_dict[playlist_url] = ""
         self.store_custom_playlists(custom_playlists_dict)
-        self.logger.log("Added custom playlist: " + playlist_url)
+        self.logger.info("Added custom playlist: " + playlist_url)
 
     def delete_custom_playlist(self, playlist_url):
         custom_playlists_dict = json.loads(self.custom_playlist_file.read_text())
         del custom_playlists_dict[playlist_url]
         self.store_custom_playlists(custom_playlists_dict)
-        self.logger.log("Removed custom playlist: " + playlist_url)
+        self.logger.info("Removed custom playlist: " + playlist_url)
 
     def delete_stored_playlist(self, playlist):
         custom_playlists_dict = json.loads(self.custom_playlist_file.read_text())
@@ -172,7 +172,7 @@ class PlaylistManager:
 
             with playlist_m3u.open("a") as append_file:
                 for track in playlist_tracks:
-                    try:
+                    try:  # TODO: I think this is fixing the exception if the track isn't in the master list, aka isn't on deezer
                         track_file_path = master_track_dict[track]["download_location"]
                         append_file.write(track_file_path + "\n")
                     except: