+++ ideemyouworthy/trackmanager.py
@@ -17,12 +17,14 @@ class TrackManager:
         if not self.master_track_file.exists():
             self.master_track_file.touch()
             self.master_track_file.write_text(json.dumps({}))
-            self.logger.log("Created master track file")
+            self.logger.info("Created master track file")
 
         self.problematic_tracks_file = Path(Path.cwd().parents[0] / "cache" / "problematic_tracks.txt")
         if not self.problematic_tracks_file.exists():
             self.problematic_tracks_file.touch()
-            self.logger.log("Created skipped local tracks file")
+            self.logger.info("Created skipped local tracks file")
+
+        self.has_finished_queue = False
 
     def get_playlist_tracks(self, playlist_id, fields):
         results = self.spotify_manager.playlist_tracks(playlist_id, fields = fields)
@@ -60,9 +62,9 @@ class TrackManager:
                 playlist_differences.pop(track, None)
 
             playlist_changes[playlist] = playlist_differences
-            self.logger.log(playlist + ": Found " + str(len(playlist_differences)) + " new tracks")
+            self.logger.info(playlist + ": Found " + str(len(playlist_differences)) + " new tracks")
 
-            playlist_file_path.write_text(json.dumps(new_playlist_songs, indent = 4))
+            playlist_file_path.write_text(json.dumps(new_playlist_songs, indent = 4, ensure_ascii = False))
 
         return playlist_changes
 
@@ -79,24 +81,25 @@ class TrackManager:
 
             tracks_to_download = tracks_to_download.union(differences)  # Checks against other lists so a new one isn't downloaded twice
 
-        self.logger.log("Cleared duplicate downloads")
+        self.logger.info("Cleared unnecessary downloads")
         return tracks_to_download
 
     def finished_queue(self, downloaded_tracks, new_playlists, playlist_changes, use_itunes):
-        self.logger.log("deezer and YouTube queues finished downloading")
+        self.has_finished_queue = True
+        self.logger.info("deezer and YouTube queues finished downloading")
 
         old_master_track_dict = json.loads(self.master_track_file.read_text())
         for track in downloaded_tracks:
             if track not in old_master_track_dict and isinstance(downloaded_tracks[track], dict):
                 old_master_track_dict[track] = downloaded_tracks[track]
 
-        self.master_track_file.write_text(json.dumps(old_master_track_dict, indent = 4))
-        self.logger.log("Saved master track list to file")
+        self.master_track_file.write_text(json.dumps(old_master_track_dict, indent = 4, ensure_ascii = False))
+        self.logger.info("Saved master track list to file")
 
         if use_itunes:
-            self.logger.log("Using iTunes")
+            self.logger.info("Using iTunes")
             itunes = win32com.client.Dispatch("iTunes.Application")
-            self.logger.log("iTunes opened")
+            self.logger.info("iTunes opened")
             itunes_sources = itunes.Sources
             itunes_playlists = None
             for source in itunes_sources:
@@ -123,12 +126,12 @@ class TrackManager:
                     if track in old_master_track_dict and isinstance(old_master_track_dict[track], dict):  # (this will be true unless there was a downloading error)
                         itunes_playlists_dict[playlist].AddFile(old_master_track_dict[track]["download_location"])
 
-            self.logger.log("Finished updating iTunes")
+            self.logger.info("Finished updating iTunes")
         else:
-            self.logger.log("Not using iTunes")
+            self.logger.info("Not using iTunes")
 
     def fix_itunes(self, itunes_playlists_dict, playlist_edits):
-        self.logger.log("Starting to fix iTunes")
+        self.logger.info("Starting to fix iTunes")
         for playlist in playlist_edits:
             extra_count = 0
             missing_count = 0
@@ -143,6 +146,7 @@ class TrackManager:
                                 extra_count += 1
                         except:
                             ""
+                            self.logger.warn("What just happened? This isn't good.")
             if playlist_edits[playlist]["missing_tracks"]:
                 for missing_track in playlist_edits[playlist]["missing_tracks"]:
                     # (technically moving right ahead without checking for the file to exist could be bad if it was deleted from the master list but not the playlist list?)
@@ -151,17 +155,17 @@ class TrackManager:
                         missing_count += 1
 
                     except:
-                        print(missing_track + " could not be added to iTunes. The max file path length on Windows is 260; the length of this file path is " + str(len(missing_track)))
+                        self.logger.warn(missing_track + " could not be added to iTunes. The max file path length on Windows is 260; the length of this file path is " + str(len(missing_track)))
                         file_path_length = str(len(missing_track))
-                        self.logger.log(missing_track + " could not be added to iTunes. File path length: " + file_path_length)
+                        self.logger.info(missing_track + " could not be added to iTunes. File path length: " + file_path_length)
                         self.store_problematic_track("(" + file_path_length + ") " + missing_track)
 
-            self.logger.log(playlist + ": Removed " + str(extra_count) + " extra tracks")
-            self.logger.log(playlist + ": Added " + str(missing_count) + " missing tracks")
+            self.logger.info(playlist + ": Removed " + str(extra_count) + " extra tracks")
+            self.logger.info(playlist + ": Added " + str(missing_count) + " missing tracks")
 
     # This verifies that the iTunes and cached playlists agree
     def verify_itunes(self):
-        self.logger.log("Starting to verify iTunes is up-to-date")
+        self.logger.info("Starting to verify iTunes is up-to-date")
         itunes = win32com.client.Dispatch("iTunes.Application")
 
         itunes_sources = itunes.Sources
@@ -183,7 +187,7 @@ class TrackManager:
         for playlist in itunes_playlists_dict:
 
             playlist_file_path = Path.cwd().parents[0] / "playlists" / (playlist + ".json")
-            official_playlist_dict = json.loads(playlist_file_path.read_text())
+            official_playlist_dict = json.loads(playlist_file_path.read_text()) # todo: this doesn't allow other playlists in itunes
             official_locations = set()
 
             for track in official_playlist_dict:
@@ -199,10 +203,10 @@ class TrackManager:
                     itunes_location = track.Location
                     itunes_locations.add(itunes_location)
                     if itunes_location not in official_locations:
-                        extra_tracks[itunes_location] = track.Name
+                        extra_tracks[itunes_location] = track.Name  # Todo: Check if there are too many or too few of a track?
 
                 except:
-                    self.logger.log("Found a non-local file. This isn't a problem but I think it might be interesting")
+                    self.logger.warn("Found a non-local file. This isn't a problem but I think it might be interesting")
 
             missing_tracks = collections.OrderedDict()
             for track in official_locations: